+++ BookService.App/Data/BookData.json
@@ -1,1 +1,1 @@
-[{"Id":1,"Name":"Harry Potter","Category":"Fiction","Author":"JK Rowling","Price":500},{"Id":2,"Name":"Sherlock Holmes","Category":"Mystery Adventure","Author":"Arthur Conan Doyle","Price":500}]
+[{"Id":1,"Name":"Harry Potter","Category":"Fiction","Author":"JK Rowling","Price":500},{"Id":3,"Name":"The Fault in our Stars","Category":"Romance","Author":"John Green","Price":350},{"Id":4,"Name":"Outliers","Category":"Business Non-Fiction","Author":"Martin Fowler","Price":500}]
